+++ ChangeLog.md
@@ -21,6 +21,7 @@
 ### Fixed
 * *Mosaic*: Possible crash in `solve_game` due to insufficient buffer size
 * *Pearl*: Fix shortcut-loop detection on squares
+* *Mines*: Fix possible hang during game generation
 
 ## 0.8.1 - 2023/12/30